requirements.txt
google-generativeai>=0.3.2
python-dotenv>=1.0.1
PyPDF2>=3.0.0
streamlit>=1.32.0
watchdog>=3.0.0
gunicorn>=21.2.0
pathlib>=1.0.1
typing>=3.7.4.3 